Moved TimeManager to data module
