+++ src/skmultiflow/utils/__init__.py
@@ -13,9 +13,8 @@ from .utils import add_dict_values
 from .utils import add_delay_to_timestamps
 from .data_structures import FastBuffer
 from .data_structures import FastComplexBuffer
-from .data_structures import TimeManager
 from ._show_versions import show_versions
 
 __all__ = ["check_random_state", "check_weights", "calculate_object_size", "get_dimensions",
            "is_scalar_nan", "normalize_values_in_dict", "FastBuffer", "FastComplexBuffer",
-           "TimeManager", "show_versions", "get_max_value_key", "add_dict_values", "add_delay_to_timestamps"]
+           "show_versions", "get_max_value_key", "add_dict_values", "add_delay_to_timestamps"]
